# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api to Ficus",
        "version": "0.1"
    },
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get all users from database",
                "description": "Get all users from database with their fields except of password",
                "operationId": "user",
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Primary key",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Name of user",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Email of user",
                                            "type": "mail"
                                        },
                                        "login": {
                                            "description": "Login of user",
                                            "type": "string"
                                        },
                                        "email_verified_at": {
                                            "description": "Time of verifiing user's mail",
                                            "type": "timestamps"
                                        },
                                        "created_at": {
                                            "description": "Time of user's registeration",
                                            "type": "timestamps"
                                        },
                                        "updated_at": {
                                            "description": "Time of last user's information changing",
                                            "type": "timestamps"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": "Example name",
                                        "email": "example@mail.com",
                                        "login": "example_user3457",
                                        "email_verified_at": null,
                                        "created_at": null,
                                        "updated_at": null
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create an user",
                "description": "Add user to database",
                "operationId": "user",
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email of user",
                                        "type": "mail"
                                    },
                                    "login": {
                                        "description": "Login of user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password of user",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Example name",
                                    "email": "example@mail.com",
                                    "login": "example_user3457",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/user/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get one user from database",
                "description": "Get one user from database with his fields except of password",
                "operationId": "user",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id of user, who we want to get",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Primary key",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Name of user",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Email of user",
                                            "type": "mail"
                                        },
                                        "login": {
                                            "description": "Login of user",
                                            "type": "string"
                                        },
                                        "email_verified_at": {
                                            "description": "Time of verifiing user's mail",
                                            "type": "timestamps"
                                        },
                                        "created_at": {
                                            "description": "Time of user's registeration",
                                            "type": "timestamps"
                                        },
                                        "updated_at": {
                                            "description": "Time of last user's information changing",
                                            "type": "timestamps"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "id": 1,
                                        "name": "Example name",
                                        "email": "example@mail.com",
                                        "login": "example_user3457",
                                        "email_verified_at": null,
                                        "created_at": null,
                                        "updated_at": null
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update one user from database",
                "description": "Update one user from database",
                "operationId": "user/{user_id}",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id of user, who we want to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of user",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Example name"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "ok"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete one user from database",
                "description": "Delete one user from database",
                "operationId": "user/{user_id}",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id of user, who we want to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "503": {
                        "description": "Service Unavailable"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "200": {
                        "description": "ok"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}